[CI] add all platforms to ci
This commits adds multiple test-platforms into CI and tests them as a third
stage.
Each platform is potentially interesting and might produce binary
artifacts in the future.
Some drive-by fixes are implemented for each platform to function.
At the moment, the docker-containers are not in the gitlab registry but
built by hand on the docker-machine.