[Fix] memory sanitizer build configuration
CMake compiles test programs to find some packages and modules,
for example `pthread`.
In order for that to work, the programs need to link properly.
Memory-Sanitizer requires the `-fsanitize=memory` flag as linker flag,
because special runtime libraries are linked.
This was missing in the normal cmake-call and lead to failure of cmake
module load.