[Refactor] feature-extraction overhauled
The driver for feature extraction is configured differently now.
Instead of accepting multiple algorithms that are used for both
detection and description it allows only ONE detector and ONE
descriptor.

Both steps are done independently. This allows mixing of different
algorithms and experimentation with algorithms that provide only one of
both steps.
It is the foundation for evaluating descriptor for the same keypoints.
Additionally that allows introducing a new filtering mechanism for
keypoints.