[Feature] implement basic match-accuracy calculation
To estimate if matched descriptors are actually the same keypoint there
needs to be done some trickery.
Because the keypoints can be projected into the camera coordinate frame
using the depth image it it is possible to check for the distance of the
keypoint.
The keypoint-pair for each match is projected into space and the
distance between these two in the same coordinate frame is measured.
This measured distance is stored.
Small distances indicate good matches, big distances indicate false
matches.

"Small" and "big" need to be defined case by case.