[Fix] hopefully fixing memory corruptions from command line
The likely for the spurious crash was, that memory was moved with the
complex initialization of the subcommands. The parser holds references
to the values that shall be written to.
These references dangle after moving, leading to memory corruption.
This dangling can be avoided when using dynamic memory for each argument set.
This way, only the POINTER to the place the arguments are moved, but not the memory
locations of the arguments themself.

Interestingly enough, the sanitizers did not catch this bug!
Fixes #47